seperated autoconsumer into own file
